Moved components to RestaurantComponents
